feat: 🎸 Revamp settlements related entities
BREAKING CHANGE: 🧨 Since some of the settlements related entities have been removed and new
ones have been added, this will require a full resync. Following are the breaking changes:
- Entities `Settlement`, `MediatorAffirmation`, `IdentityInstructions` have been removed
- Changes in `Venue` entity
  1. New attribute `signers` has been added to track all the allowed signers associated with them
  2. New derived field `instructions: [Instruction]!` has been added to track instruction info
   associated a Venue
- Adds new enum `InstructionTypeEnum` to track instruction type
- Changes in `Instruction` entity
  1. `eventId`, `eventIdx` have been removed
  2. `settlementType: String!` has been modified to `type: InstructionTypeEnum!`
  3. `endAfterBlock: Int` has been added to track block after which `SettleManual` type
  instructions can be executed
  4. `mediators: [String!]!` has been added to track list of mediators for an instruction
  5. New derived fields namely - `parties: [InstructionParty]!`,
  `affirmations: [InstructionAffirmation]!`, `events: [InstructionEvent]!` have been added to
  query more information about an instruction
- Changes in `Leg` entity
  1. Attribute `settlement: Settlement` has been removed
  2. Attribute `instruction` is now non-null
  3. The `from` and `to` part of the legs have been split separately into identity and
  portfolio number to support off chain legs. The new attributes are
  `from: String!`, `fromPortfolio: number`, `to: String!`, `toPortfolio: number`
- Following new entities have been added:
  1. `InstructionParty` - to track all the DIDs associated with an instruction.
  2. `InstructionAffirmation`- to track all the affirmations received for an instruction
  3. `OffChainReceipt` - to track all the receipts used in affirming off chain legs
  4. `InstructionEvent` - to track all the event associated with an instruction.
  This will give a clear picture about the lifetime of an instruction on how it got created,
   to affirming, to execution
