+++ src/mappings/entities/index.ts
@@ -19,7 +19,8 @@ export * from './mapNfts';
 export * from './mapPolyxTransaction';
 export * from './mapPortfolio';
 export * from './mapProposal';
-export * from './mapSettlement';
+export * from './settlements/mapSettlement';
+export * from './settlements/mapVenue';
 export * from './mapStakingEvent';
 export * from './mapStatistics';
 export * from './mapSto';
