+++ src/utils/accounts.ts
@@ -1,0 +1,16 @@
+import { decodeAddress, encodeAddress } from '@polkadot/keyring';
+import { Codec } from '@polkadot/types/types';
+import { u8aToHex } from '@polkadot/util';
+
+export const serializeAccount = (item: Codec): string | undefined => {
+  const s = item.toString();
+
+  if (s.trim().length === 0) {
+    return undefined;
+  }
+  return u8aToHex(decodeAddress(item.toString(), false, item.registry.chainSS58));
+};
+
+export const getAccountKey = (item: string, ss58Format?: number): string => {
+  return encodeAddress(item.toString(), ss58Format);
+};
