+++ tsconfig.json
@@ -11,11 +11,11 @@
     "outDir": "dist",
     "baseUrl": ".",
     "rootDir": "src",
-    "target": "es2017",
-    "paths": {
-      // this is the package name we use (in the interface imports, --package for generators)
-      "polymesh-subql/*": ["src/*"],
-    }
+    "target": "es2017"
   },
-  "include": ["src/**/*", "node_modules/@subql/types/dist/global.d.ts", "node_modules/@subql/types-core/dist/global.d.ts"]
+  "include": [
+    "src/**/*",
+    "node_modules/@subql/types/dist/global.d.ts",
+    "node_modules/@subql/types-core/dist/global.d.ts"
+  ]
 }
